Use robolectric for the map test by moving the simple provider to the src folder
