+++ app/src/main/java/com/github/swent/echo/compose/map/SimpleMapViewProvider.kt
@@ -1,11 +1,10 @@
-package com.github.swent.echo.di
+package com.github.swent.echo.compose.map
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
-import com.github.swent.echo.compose.map.IMapViewProvider
 import com.github.swent.echo.data.model.Event
 
 /**
